Peak detector sync finished and debugged; preferences added
Peak detector now has two parameters: absolute threshold and standard
deviation threshold.  Both have to be cleared for the maximum to register.
After the thresholds are cleared, the system returns the timestamp of maximum
acceleration seen after the threshold.  Experiments show that bumping the
phones together provides a very poor reference; shaking or rotation should be
used instead.  Default values for the thresholds are still optimized for the
bumping approach.