GraphView shows the running acceleration trace; NTP sync saves data.
Acceleration magnitude trace is shown in GraphView; drawing of the trace is
done by a HandlerThread that receives messages from DataLogger (currently
they are passed by the acceleration even listener).

NTP sync class now saves the timestamp and reference offset in
SharedPreferences (filename "datastorePrefs"). If NTP lookup fails and
there's no instance of ntpClient with good data in the current application,
the saved values are used (provided the phone hasn't been rebooted after
they've been saved).

File list display view is now invoked through a menu item and splash screen
is shown at application start and resume.

Strange behavior in the handleMessage method of GraphView: when resuming the
application with accelerometers running, it seems to receive a message with a
null object field, necessitating explicit checking for null.

Strange behavior of the GraphView thread: upon stopping, we get exceptions
due to messages are being sent to a dead thread. The call to to stop logging
and unregister sensor listeners comes before that to quit the handler thread;
it's not clear to me yet why the execution order gets messed up.